stole HTTP::Daemon from perl6-examples
mberends++. HTTP::Daemon now serves the Lobster in the browser, without any
modifications.

Run the following command to get the Lobster in your browser:

perl6 bin/run-lobster

Flipping it works, but crashing doesn't work as in Rack -- that is, it doesn't
show the exception in the browser, it just crashes and doesn't come back. Need
to port Rack::ShowExceptions to make this work.