[HTTP::Daemon] removed from repo, added dep
Ideally, Web.pm should be completely webserver-agnostic, but right now
HTTP::Daemon is our only realistic option, so we'll go with that. In
Ruby's Rack, the Lobster example uses WEBrick per default; they can
assume that it's already installed because it ships with Ruby. In Perl 6,
we have no corresponding luxury, yet.